Updated project to 1.1b61

+++ src/main/java/de/mendelson/comm/as2/timing/MessageDeleteController.java
@@ -1,4 +1,4 @@
-//$Header: /as2/de/mendelson/comm/as2/timing/MessageDeleteController.java 36    24.04.20 9:52 Heller $
+//$Header: /as2/de/mendelson/comm/as2/timing/MessageDeleteController.java 45    27/01/22 11:34 Heller $
 package de.mendelson.comm.as2.timing;
 
 import de.mendelson.comm.as2.clientserver.message.RefreshClientMessageOverviewList;
@@ -10,7 +10,9 @@ import de.mendelson.comm.as2.preferences.PreferencesAS2;
 import de.mendelson.comm.as2.preferences.ResourceBundlePreferences;
 import de.mendelson.comm.as2.server.AS2Server;
 import de.mendelson.util.MecResourceBundle;
+import de.mendelson.util.NamedThreadFactory;
 import de.mendelson.util.clientserver.ClientServer;
+import de.mendelson.util.database.IDBDriverManager;
 import de.mendelson.util.systemevents.SystemEvent;
 import de.mendelson.util.systemevents.SystemEventManagerImplAS2;
 import java.io.File;
@@ -18,6 +20,7 @@ import java.nio.file.Files;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Paths;
 import java.sql.Connection;
+import java.sql.Statement;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -26,6 +29,7 @@ import java.util.List;
 import java.util.MissingResourceException;
 import java.util.ResourceBundle;
 import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
 
@@ -40,7 +44,7 @@ import java.util.logging.Logger;
  * Controls the timed deletion of AS2 entries from the log
  *
  * @author S.Heller
- * @version $Revision: 36 $
+ * @version $Revision: 45 $
  */
 public class MessageDeleteController {
 
@@ -50,17 +54,25 @@ public class MessageDeleteController {
     private Logger logger = Logger.getLogger(AS2Server.SERVER_LOGGER_NAME);
     private PreferencesAS2 preferences = new PreferencesAS2();
     private MessageDeleteThread deleteThread;
+    private final ScheduledExecutorService scheduledExecutor = Executors.newSingleThreadScheduledExecutor(
+        new NamedThreadFactory("old-transactions-housekeeping"));
     private ClientServer clientserver = null;
     private MecResourceBundle rb = null;
     private MecResourceBundle rbTime = null;
     private Connection configConnection;
     private Connection runtimeConnection;
+    private LogAccessDB logAccess;
+    private IDBDriverManager dbDriverManager;
+    private MessageAccessDB messageAccess;
 
-    public MessageDeleteController(ClientServer clientserver, Connection configConnection,
+    public MessageDeleteController(ClientServer clientserver, IDBDriverManager dbDriverManager, Connection configConnection,
             Connection runtimeConnection) {
         this.clientserver = clientserver;
         this.configConnection = configConnection;
         this.runtimeConnection = runtimeConnection;
+        this.dbDriverManager = dbDriverManager;
+        this.logAccess = new LogAccessDB(dbDriverManager);
+        this.messageAccess = new MessageAccessDB(this.dbDriverManager, this.configConnection, this.runtimeConnection);
         //Load default resourcebundle
         try {
             this.rb = (MecResourceBundle) ResourceBundle.getBundle(
@@ -83,20 +95,27 @@ public class MessageDeleteController {
      */
     public void startAutoDeleteControl() {
         this.deleteThread = new MessageDeleteThread(this.configConnection, this.runtimeConnection);
-        Executors.newSingleThreadExecutor().submit(this.deleteThread);
+        this.scheduledExecutor.scheduleWithFixedDelay(this.deleteThread, 1, 1, TimeUnit.MINUTES);
     }
 
     /**
-     * Deletes a message entry from the log. Clears all files
+     * Deletes a sub list of the whole delete order from the log
+     *
+     * @param infoList
+     * @param broadcastRefresh
+     * @param deleteLog
      */
-    public void deleteMessageFromLog(AS2MessageInfo info, boolean broadcastRefresh, StringBuilder deleteLog) {
-        LogAccessDB logAccess = new LogAccessDB(this.configConnection, this.runtimeConnection);
-        logAccess.deleteMessageLog(info.getMessageId());
-        MessageAccessDB messageAccess = new MessageAccessDB(this.configConnection, this.runtimeConnection);
+    private void deleteMessageFromLogPart(List<AS2MessageInfo> infoList, boolean broadcastRefresh, StringBuilder deleteLog,
+            Connection runtimeConnectionNoAutoCommit) {
         try {
+            List<String> messageIdList = new ArrayList<String>();
+            for (AS2MessageInfo messageInfo : infoList) {
+                messageIdList.add(messageInfo.getMessageId());
+            }
+            this.logAccess.deleteMessageLog(messageIdList, runtimeConnectionNoAutoCommit);
             //delete all raw files from the disk
-            List<String> rawfilenames = messageAccess.getRawFilenamesToDelete(info);
-            if (rawfilenames != null) {
+            List<String> rawfilenames = this.messageAccess.getRawFilenamesToDelete(messageIdList, runtimeConnectionNoAutoCommit);
+            if (rawfilenames != null && !rawfilenames.isEmpty()) {
                 for (String rawfilename : rawfilenames) {
                     try {
                         Files.delete(Paths.get(rawfilename));
@@ -115,7 +134,7 @@ public class MessageDeleteController {
                     deleteLog.append(System.lineSeparator());
                 }
             }
-            messageAccess.deleteMessage(info);
+            this.messageAccess.deleteMessages(messageIdList, runtimeConnectionNoAutoCommit);
         } catch (Exception e) {
             this.logger.severe("deleteMessageFromLog: " + e.getMessage());
             SystemEventManagerImplAS2.systemFailure(e, SystemEvent.TYPE_PROCESSING_ANY);
@@ -125,11 +144,73 @@ public class MessageDeleteController {
         }
     }
 
+    /**
+     * Deletes a message entry from the log. Clears all files. Main entry point
+     * for a delete operation
+     */
+    public void deleteMessagesFromLog(List<AS2MessageInfo> infoList, boolean broadcastRefresh, StringBuilder deleteLog) {
+        int bulkSize = 10;
+        //a new connection to the database is required because the message storage contains several tables and all this has to be transactional
+        Connection runtimeConnectionNoAutoCommit = null;
+        String transactionname = "Message_deleteFromLog";
+        Statement transactionStatement = null;
+        try {
+            runtimeConnectionNoAutoCommit = this.dbDriverManager.getConnectionWithoutErrorHandling(IDBDriverManager.DB_RUNTIME);
+            runtimeConnectionNoAutoCommit.setAutoCommit(false);
+            transactionStatement = runtimeConnectionNoAutoCommit.createStatement();
+            //start transaction
+            this.dbDriverManager.startTransaction(transactionStatement, transactionname);
+            //lock tables for delete
+            this.dbDriverManager.setTableLockDELETE(transactionStatement,
+                    new String[]{
+                        "messages",
+                        "messagelog",
+                        "mdn",
+                        "payload"
+                    });
+            List<AS2MessageInfo> subList = new ArrayList<AS2MessageInfo>();
+            for (int i = 0; i < infoList.size(); i++) {
+                subList.add(infoList.get(i));
+                if (i != 0 && i % bulkSize == 0) {
+                    this.deleteMessageFromLogPart(subList, broadcastRefresh, deleteLog, runtimeConnectionNoAutoCommit);
+                    subList.clear();
+                }
+            }
+            if (!subList.isEmpty()) {
+                this.deleteMessageFromLogPart(subList, broadcastRefresh, deleteLog, runtimeConnectionNoAutoCommit);
+            }
+            //all ok - commit
+            this.dbDriverManager.commitTransaction(transactionStatement, transactionname);
+        } catch (Exception e) {
+            try {
+                //an error occured - rollback transaction and release all table locks
+                this.dbDriverManager.rollbackTransaction(transactionStatement);
+            } catch (Exception ex) {
+                SystemEventManagerImplAS2.systemFailure(ex, SystemEvent.TYPE_DATABASE_ANY);
+            }
+            this.logger.severe("MessageAccessDB.deleteMessagesFromLog: " + e.getMessage());
+            SystemEventManagerImplAS2.systemFailure(e, SystemEvent.TYPE_DATABASE_ANY);
+        } finally {
+            if (transactionStatement != null) {
+                try {
+                    transactionStatement.close();
+                } catch (Exception e) {
+                    //nop
+                }
+            }
+            if (runtimeConnectionNoAutoCommit != null) {
+                try {
+                    runtimeConnectionNoAutoCommit.close();
+                } catch (Exception ex) {
+                    //nop
+                }
+            }
+        }
+
+    }
+
     public class MessageDeleteThread implements Runnable {
 
-        private boolean stopRequested = false;
-        //wait this time between checks
-        private final long WAIT_TIME_ONE_MINUTE = TimeUnit.MINUTES.toMillis(1);
         //DB connection
         private Connection configConnection;
         private Connection runtimeConnection;
@@ -141,26 +222,18 @@ public class MessageDeleteController {
 
         @Override
         public void run() {
-            Thread.currentThread().setName("Contol auto AS2 message delete");
-            while (!stopRequested) {
                 try {
-                    try {
-                        Thread.sleep(WAIT_TIME_ONE_MINUTE);
-                    } catch (InterruptedException e) {
-                        //nop
-                    }
                     if (preferences.getBoolean(PreferencesAS2.AUTO_MSG_DELETE)) {
                         MessageAccessDB messageAccess = null;
                         try {
-                            messageAccess = new MessageAccessDB(this.configConnection, this.runtimeConnection);
+                        messageAccess = new MessageAccessDB(dbDriverManager, this.configConnection, this.runtimeConnection);
                             long olderThan = System.currentTimeMillis()
                                     - TimeUnit.SECONDS.toMillis(
                                             preferences.getInt(PreferencesAS2.AUTO_MSG_DELETE_OLDERTHAN)
                                             * preferences.getInt(PreferencesAS2.AUTO_MSG_DELETE_OLDERTHAN_MULTIPLIER_S));
                             List<AS2MessageInfo> overviewList = messageAccess.getMessagesOlderThan(olderThan, -1);
                             if (overviewList != null) {
-                                List<AS2MessageInfo> deletedTransactionList = new ArrayList<AS2MessageInfo>();
-                                List<StringBuilder> deletedTransactionLog = new ArrayList<StringBuilder>();
+                            List<AS2MessageInfo> deleteList = new ArrayList<AS2MessageInfo>();
                                 for (AS2MessageInfo messageInfo : overviewList) {
                                     if (messageInfo.getState() == AS2Message.STATE_FINISHED || messageInfo.getState() == AS2Message.STATE_STOPPED) {
                                         if (preferences.getBoolean(PreferencesAS2.AUTO_MSG_DELETE_LOG)) {
@@ -179,15 +252,14 @@ public class MessageDeleteController {
                                                         timeUnit
                                                     }));
                                         }
-                                        StringBuilder singleDeleteLog = new StringBuilder();
-                                        deleteMessageFromLog(messageInfo, true, singleDeleteLog);
-                                        deletedTransactionLog.add(singleDeleteLog);
-                                        deletedTransactionList.add(messageInfo);
+                                    deleteList.add(messageInfo);
                                     }
                                 }
+                            StringBuilder deletedTransactionLog = new StringBuilder();
+                            deleteMessagesFromLog(deleteList, true, deletedTransactionLog);
                                 //fire system event if automatic log deletes should be loged
                                 if (preferences.getBoolean(PreferencesAS2.AUTO_MSG_DELETE_LOG)) {
-                                    this.fireSystemEventTransactionsDeletedByMaintenance(olderThan, deletedTransactionList,
+                                this.fireSystemEventTransactionsDeletedByMaintenance(olderThan, deleteList,
                                             deletedTransactionLog);
                                 }
                             }
@@ -202,16 +274,15 @@ public class MessageDeleteController {
                     SystemEventManagerImplAS2.systemFailure(e, SystemEvent.TYPE_PROCESSING_ANY);
                 }
             }
-        }
 
         /**
          * Fire a system event that the system maintenance process has deleted
          * transactions
          */
         private void fireSystemEventTransactionsDeletedByMaintenance(long olderThan, List<AS2MessageInfo> deletedTransactionList,
-                List<StringBuilder> singleTransactionDeleteLog) {
+                StringBuilder transactionDeleteLog) {
             //Do not fire an event if there are no deleted transactions
-            if( deletedTransactionList.isEmpty()){
+            if (deletedTransactionList.isEmpty()) {
                 return;
             }
             DateFormat dateFormat = SimpleDateFormat.getDateTimeInstance(DateFormat.SHORT, DateFormat.MEDIUM);
@@ -225,8 +296,6 @@ public class MessageDeleteController {
             builder.append(System.lineSeparator()).append(System.lineSeparator());
             for (int i = 0; i < deletedTransactionList.size(); i++) {
                 AS2MessageInfo singleInfo = deletedTransactionList.get(i);
-                StringBuilder singleDeleteLog = singleTransactionDeleteLog.get(i);
-                builder.append("---").append(System.lineSeparator());
                 builder.append("[");
                 builder.append(rb.getResourceString("transaction.deleted.transactiondate",
                         dateFormat.format(singleInfo.getInitDate())));
@@ -237,9 +306,9 @@ public class MessageDeleteController {
                 builder.append(") ");
                 builder.append(singleInfo.getMessageId());
                 builder.append(System.lineSeparator());
-                builder.append(singleDeleteLog);
-                builder.append(System.lineSeparator());
             }
+            builder.append("---").append(System.lineSeparator());
+            builder.append(transactionDeleteLog);
             event.setBody(builder.toString());
             SystemEventManagerImplAS2.newEvent(event);
         }
